INNOVAPLUS/INNOVA/requirements.txt
fastapi
uvicorn[standard]
pydantic
llama-cpp-python
python-dotenv

